Safari-609.1.11.1/LayoutTests/js/dom/Promise-onFulfilled-deep-expected.txt

Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is undefined
PASS result is 42
PASS successfullyParsed is true

TEST COMPLETE
